autocast settings from variations to float

+++ comsol.py
@@ -95,7 +95,7 @@ def read_csv(path, unpack_settings = True, explicit_settings = True, clean_names
         
         # merging does not work! all values are the same, wtf....
         for c in sdf.columns:
-            df2[c] = sdf[c].values
+            df2[c] = sdf[c].values.astype(float)
     
     
     return(df2)
